Updated code to have csv writing
